Restyles Clear Data button to match the Get Data Button
Adds Clear Data functionality to button
