+++ src/components/App.scss
@@ -65,35 +65,26 @@
       }
     }
 
-      .clear-data-button {
-        border: none;
-        color: #72bfca;
-        background-color: white;
-        cursor: pointer;
-        font-family: 'Montserrat', sans-serif;
-        padding: 0px;
+    button {
+      width: 82px;
+      height: 31px;
+      padding: 0 2px;
+      margin-left: 13px;
+      border-radius: 3px;
+      border: solid 1px #000;
+      background-color: #fff;
+      color: #000;
+      font-family: "Montserrat", sans-serif;
+      cursor: pointer;
+      &:hover {
+        background-color: rgba(198, 198, 198, 0.25);
       }
-      .get-data-button {
-        width: 82px;
-        height: 31px;
-        margin-left: 13px;
-        border-radius: 3px;
-        border: solid 1px #000;
-        background-color: #fff;
-        color: #000;
-        font-family: 'Montserrat', sans-serif;
-        cursor: pointer;
-        &:hover {
-          background-color: rgba(198, 198, 198, 0.25);
-        }
-        &:disabled {
-          cursor: default;
-          color: #a2a2a2;
-          border: solid 1px #a2a2a2;
-        }
+      &:disabled {
+        cursor: default;
+        color: #a2a2a2;
+        border: solid 1px #a2a2a2;
       }
-
-
+    }
   }
 }
